// apps/demos/Demos/Common/PopupAndNotificationsOverview/React/types.ts
export type HouseType = {
  ID: string;
  Favorite: boolean;
  Address: string;
  City: string;
  State: string;
  ZipCode: string;
  Beds: string;
  Baths: string;
  HouseSize: number;
  LotSize: string;
  Price: number;
  Coordinates: string;
  Features: string;
  YearBuilt: string;
  Type: string;
  Status: string;
  Image: string;
  Agent: {
    Name: string;
    Phone: string;
    Picture: string;
  };
};